Work on Markov weather generator functionality
- added second example to inst/extdata (because turning it on during
runtime does currently not work)
- added example to sw_exec
- fixed methods for class ‘swWeather’ (they used an incorrect slot name)
- NOTE: SOILWAT function onGetInputDataFromFiles() reads the
mkv_prob.in and mkv_covar.in files only if the Markov flag is set. That
is if the flag is unset, but later set during runtime, e.g.,
swWeather_UseMarkov(sw_in) <- TRUE, then the simulation will fail
because no markov parameters are present. I suggest that the code
should read the markov input files whenever they are present — and also
provide a end-user function in the Rsoilwat31 package to request that
these files be read in during runtime.