Function 'dbW_upgrade_to_rSOILWAT2' gained 'with_resume' and 'clean_cache'
- Function 'dbW_upgrade_to_rSOILWAT2' gained two arguments
’with_resume' and 'clean_cache’. This functionality is useful, for
instance, in cases when weather database files with many records are
upgraded, but the process had to be interrupted.

- with_resume: If TRUE, then digest of call and progress status are
stored in a (hidden) temporary file on disk on the same path as
dbWeatherDataFile. If such a temporary file is present, then upgrade is
resumed from where progress status indicates. Otherwise, ignore
temporary disk files.

- clean_cache: If TRUE, then temporary file generated with argument
‘with_resume’ set to TRUE, if present, is removed once upgrade
completed successfully.